requirements.txt
numpy
pandas
python-dotenv
pymongo
certifi
pymongo[srv]==3.12
scikit-learn

# -e .